# Add locking to loki.write sink + nits