--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -130,8 +130,6 @@ Main (unreleased)
 - Mark `password` argument of `loki.source.kafka` as a `secret` rather than a `string`. (@harsiddhdave44)
 
 - Fixed a bug where UDP syslog messages were never processed (@joshuapare)
-  "blgithub.com/grafana/loki/pkg/logproto"
-  loki_util "github.com/grafana/loki/pkg/util"
 
 - Updating configuration for `loki.write` no longer drops data. (@thepalbi)