--- a/component/loki/write/write_test.go
+++ b/component/loki/write/write_test.go
@@ -3,6 +3,7 @@ package write
 import (
 	"context"
 	"fmt"
+	"go.uber.org/atomic"
 	"math"
 	"net/http"
 	"net/http/httptest"
@@ -302,5 +303,100 @@ func testMultipleEndpoint(t *testing.T, alterArgs func(arguments *Arguments)) {
 			require.Len(t, req.Streams[0].Entries, 1)
 			require.Equal(t, req.Streams[0].Entries[0].Line, "writing some text")
 		}
+	}
+}
+
+type testCase struct {
+	linesCount  int
+	seriesCount int
+}
+
+func BenchmarkLokiWrite(b *testing.B) {
+	for name, tc := range map[string]testCase{
+		"100 lines, single series": {
+			linesCount:  100,
+			seriesCount: 1,
+		},
+		"100k lines, 100 series": {
+			linesCount:  100_000,
+			seriesCount: 100,
+		},
+	} {
+		b.Run(name, func(b *testing.B) {
+			benchSingleEndpoint(b, tc, func(arguments *Arguments) {})
+		})
+	}
+}
+
+func benchSingleEndpoint(b *testing.B, tc testCase, alterConfig func(arguments *Arguments)) {
+	// Set up the server that will receive the log entry, and expose it on ch.
+	var seenLines atomic.Int64
+	ch := make(chan logproto.PushRequest)
+
+	// just count seenLines for each entry received
+	go func() {
+		for pr := range ch {
+			count := 0
+			for _, str := range pr.Streams {
+				count += len(str.Entries)
+			}
+			seenLines.Add(int64(count))
+		}
+	}()
+
+	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		var pushReq logproto.PushRequest
+		err := loki_util.ParseProtoReader(context.Background(), r.Body, int(r.ContentLength), math.MaxInt32, &pushReq, loki_util.RawSnappy)
+		if err != nil {
+			w.WriteHeader(http.StatusBadRequest)
+			return
+		}
+		tenantHeader := r.Header.Get("X-Scope-OrgID")
+		require.Equal(b, tenantHeader, "tenant-1")
+
+		ch <- pushReq
+	}))
+	defer srv.Close()
+
+	// Set up the component Arguments.
+	cfg := fmt.Sprintf(`
+		endpoint {
+			url        = "%s"
+			batch_wait = "10ms"
+			tenant_id  = "tenant-1"
+		}
+	`, srv.URL)
+	var args Arguments
+	require.NoError(b, river.Unmarshal([]byte(cfg), &args))
+
+	alterConfig(&args)
+
+	// Set up and start the component.
+	testComp, err := componenttest.NewControllerFromID(util.TestLogger(b), "loki.write")
+	require.NoError(b, err)
+	go func() {
+		err = testComp.Run(componenttest.TestContext(b), args)
+		require.NoError(b, err)
+	}()
+	require.NoError(b, testComp.WaitExports(time.Second))
+
+	// get exports from component
+	exports := testComp.Exports().(Exports)
+
+	for i := 0; i < b.N; i++ {
+		for j := 0; j < tc.linesCount; j++ {
+			logEntry := loki.Entry{
+				Labels: model.LabelSet{"foo": model.LabelValue(fmt.Sprintf("bar-%d", i%tc.seriesCount))},
+				Entry: logproto.Entry{
+					Timestamp: time.Now(),
+					Line:      "very important log",
+				},
+			}
+			exports.Receiver.Chan() <- logEntry
+		}
+
+		require.Eventually(b, func() bool {
+			return int64(tc.linesCount) == seenLines.Load()
+		}, time.Minute, time.Second, "haven't seen expected number of lines")
 	}
 }