--- a/pkg/flow/componenttest/context.go
+++ b/pkg/flow/componenttest/context.go
@@ -6,7 +6,7 @@ import (
 )
 
 // TestContext returns a context which cancels itself when t finishes.
-func TestContext(t *testing.T) context.Context {
+func TestContext(t testing.TB) context.Context {
 	ctx, cancel := context.WithCancel(context.Background())
 	t.Cleanup(cancel)
 	return ctx