… component/common/loki/client/client_writeto.go
… component/common/loki/client/client_writeto_test.go