#### PR Description
This PR first addresses some small nits pending from https://github.com/grafana/agent/pull/5590. Also, adds a locking guard over the sink channel that lives inside `loki.write`. Both the Update and Run functions of that component interact with the `sink` entry handler, used to send received entries for remote write. There was no locking on the `Run` sink use. This PR adds it, and adds a bench to validate there's no significant change, assumming `Updates` calls don't occur often.

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/agent/component/loki/write
                                      │ /Users/pablo/work/benchmarks/loki_write_no_lock.txt │ /Users/pablo/work/benchmarks/loki_write_rw_lock.txt │
                                      │                       sec/op                        │            sec/op              vs base              │
LokiWrite/100_lines,_single_series-10                                            1.003 ± 1%                      1.003 ± 0%       ~ (p=0.937 n=6)
LokiWrite/100k_lines,_100_series-10                                              1.169 ± 1%                      1.182 ± 1%       ~ (p=0.394 n=6)
geomean                                                                          1.083                           1.089       +0.56%

                                      │ /Users/pablo/work/benchmarks/loki_write_no_lock.txt │ /Users/pablo/work/benchmarks/loki_write_rw_lock.txt │
                                      │                        B/op                         │             B/op               vs base              │
LokiWrite/100_lines,_single_series-10                                         214.4Ki ± 20%                   220.2Ki ± 16%       ~ (p=0.240 n=6)
LokiWrite/100k_lines,_100_series-10                                           107.7Mi ±  3%                   107.0Mi ±  3%       ~ (p=0.485 n=6)
geomean                                                                       4.748Mi                         4.797Mi        +1.01%

                                      │ /Users/pablo/work/benchmarks/loki_write_no_lock.txt │ /Users/pablo/work/benchmarks/loki_write_rw_lock.txt │
                                      │                      allocs/op                      │           allocs/op            vs base              │
LokiWrite/100_lines,_single_series-10                                          2.010k ± 17%                    2.007k ± 17%       ~ (p=0.853 n=6)
LokiWrite/100k_lines,_100_series-10                                            703.0k ±  0%                    703.1k ±  0%       ~ (p=0.310 n=6)
geomean                                                                        37.59k                          37.57k        -0.07%
```

Also, this PR removes the client write to and adds a missing CHANGELOG entry.

#### Which issue(s) this PR fixes

<!-- Uncomment the following line if you want that GitHub issue gets automatically closed after merging the PR -->
<!-- Fixes #issue_id -->

#### Notes to the Reviewer

#### PR Checklist

<!-- Remove items that do not apply. For completed items, change [ ] to [x]. -->

- [ ] CHANGELOG.md updated
- [ ] Documentation added
- [ ] Tests updated
- [ ] Config converters updated